move fuse fn in fuse.rs
