+++ src/channel.rs
@@ -27,136 +27,7 @@ pub struct Channel {
     fd: c_int,
 }
 
-use libc::getuid;
-use libc::getgid;
-use libc::mount;
-use std::fs::OpenOptions;
-use std::os::unix::io::AsRawFd;
-use std::os::unix::io::IntoRawFd;
-
-fn fuse_mount_sys(mountpoint: &PathBuf, flags: u64) -> i32
-{
-    // TODO:Check args
-    // TODO:Check mountpoint
-    // TODO:Check nonempty
-    // TODO:Check auto_umount
-    let f = OpenOptions::new().read(true).write(true).open("/dev/fuse").unwrap();
-
-
-    // TODO:Check f
-    // from:sdcard.c    sprintf(opts, "fd=%i,rootmode=40000,default_permissions,allow_other,"
-    //                                "user_id=%d,group_id=%d", fd, uid, gid);
-    let opts = format!("fd={},rootmode={},default_permissions,allow_other,user_id={},group_id={}",
-                       f.as_raw_fd(),40000,
-                       unsafe{getuid()}, unsafe{getgid()});
-    // TODO: Add kernel opt
-    // 
-    //TODO: understand:
-    /*
-	strcpy(type, mo->blkdev ? "fuseblk" : "fuse");
-	if (mo->subtype) {
-		strcat(type, ".");
-		strcat(type, mo->subtype);
-	}
-	strcpy(source,
-	       mo->fsname ? mo->fsname : (mo->subtype ? mo->subtype : devname));
-    */
-    info!("{}", opts);
-    let c_sources = CString::new("/dev/fuse").unwrap();
-    let c_mnt = CString::new(mountpoint.as_os_str().as_bytes()).unwrap();
-    let c_fs = CString::new("fuse").unwrap();
-    let c_opts = CString::new(opts).unwrap();
-    info!("MOUNT({:?} {:?} {:?} {:?})", c_sources, c_mnt, c_fs, c_opts);
-    let res = unsafe{
-        #[cfg(target_pointer_width = "32")]
-        let flags = flags as u32;
-        mount(c_sources.as_ptr(), c_mnt.as_ptr(), c_fs.as_ptr(), flags, c_opts.as_ptr() as *mut c_void)
-    };
-    if res < 0 {
-        return res;
-    } else {
-        return f.into_raw_fd();
-    }
-}
-
-// /usr/include/sys/mount.h
-const MS_NOSUID :u64 = 2;
-const MS_NODEV  :u64 = 4;
-const FUSE_COMMFD_ENV: &str = "_FUSE_COMMFD";
-
-use sendfd::UnixSendFd;
-use std::process::{Command, Stdio};
-use std::os::unix::net::UnixStream;
-
-fn fuse_mount_fusermount(mountpoint: &PathBuf, _: &fuse_args) -> i32
-{
-    let (sock1, sock2) = match UnixStream::pair() {
-        Ok(res) => res,
-        Err(err) => {
-            error!("{:?}", err);
-            return -1;
-        }
-    };
-    let fd = sock2.as_raw_fd();
-
-    unsafe {
-        libc::fcntl(fd, libc::F_SETFD, 0);
-    }
-    match Command::new("fusermount")
-        .arg("-o")
-        .arg("rw,nosuid,nodev,allow_other,nonempty")
-        .arg("--")
-        .arg(mountpoint)
-        .env(FUSE_COMMFD_ENV, format!("{}", fd))
-        .stdout(Stdio::inherit())
-        .spawn() {
-            Ok(_) => (),
-            Err(err) => {
-                error!("{:?}", err);
-                return -1;
-            }
-        };
-
-    return sock1.recvfd().unwrap_or(-1);
-}
-
-use errno::errno;
-
-fn fuse_kern_mount(mountpoint: &PathBuf, args: &fuse_args) -> i32
-{
-    let flags = MS_NOSUID | MS_NODEV;
-
-    // TODO: parse fuse_args
-    /*
-       pub argc: c_int,
-       pub argv: *const *const c_char,
-       pub allocated: c_int,
-       */
-    // TODO: check if allow_other and allow_root aren't mutually active
-    // TODO: check if help
-    // TODO: get kernel/other flags options
-
-    let mut res = fuse_mount_sys(mountpoint, flags);
-    if res < 0 {
-        let err = errno().0;
-        error!("fuse_mount_sys errno: {}", err);
-        // TODO: error
-        if err == libc::EPERM {
-            warn!("fuse_mount_sys EPERM: backing to fusermount...");
-            res = fuse_mount_fusermount(mountpoint, args);
-        } else {
-            panic!("Err {}: fuse_kern_mount panic !", err);
-        }
-    }
-    println!("fantafs: fuse_mount_compat25: fd={}", res);
-    // TODO: ERROR
-    res
-}
-
-fn fuse_mount_compat25(mountpoint: &PathBuf, args: &fuse_args) -> i32
-{
-    fuse_kern_mount(mountpoint, args)
-}
+use fuse::fuse_mount_compat25;
 
 impl Channel {
     /// Create a new communication channel to the kernel driver by mounting the
